requirements.txt
streamlit
pytesseract
Pillow
pdf2image
opencv-python
numpy
langdetect
googletrans==4.0.0-rc1
google-generativeai
faiss-cpu
python-dotenv

# Pinned versions to resolve langchain conflicts and network issues
httpx
httpcore
langchain
langchain-community
langchain-google-genai